% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{clear_console}
\alias{clear_console}
\title{Clear the console}
\usage{
clear_console()
}
\value{
Invisibly returns NULL
}
\description{
This can be handy to suppress messages at startup even when R has been
started without the --quiet flag
}
