% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{new_quarto_post}
\alias{new_quarto_post}
\title{Create a blank quarto blog post}
\usage{
new_quarto_post(slug, date = as.character(Sys.Date()), projdir = ".")
}
\arguments{
\item{slug}{Slug for the post}

\item{date}{Date for the post}

\item{projdir}{Project directory for the blog}
}
\value{
Invisibly returns NULL
}
\description{
Create a blank quarto blog post
}
