% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{open_folder}
\alias{open_folder}
\title{Switch to a new folder using its nickname}
\usage{
open_folder(id, folders = getOption("etal.folders"), new_session = TRUE)
}
\arguments{
\item{folders}{Named vector of folder paths}

\item{new_session}{Should a new session be started? (default = TRUE)}

\item{name}{Nickname used to refer to the folder}
}
\description{
Switch to a new folder using its nickname
}
