% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{rstudio_dark}
\alias{rstudio_dark}
\alias{rstudiothemes}
\alias{rstudio_light}
\alias{rstudio_bright}
\title{Switch between RStudio themes using options}
\usage{
rstudio_dark()

rstudio_light()

rstudio_bright()
}
\value{
As \code{rstudioapi::applyTheme()}
}
\description{
Switch between RStudio themes using options
}
