% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{session_status}
\alias{session_status}
\title{Concisely report session status information}
\usage{
session_status()
}
\value{
Invisibly returns NULL
}
\description{
Concisely report session status information
}
